Move NotFoundError to postgres directory.
