--- services/app-api/src/errors/errors.ts
@@ -1,9 +1,0 @@
-class NotFoundError extends Error {
-    constructor(message: string) {
-        super(message)
-
-        Object.setPrototypeOf(this, NotFoundError.prototype)
-    }
-}
-
-export { NotFoundError }
