--- services/app-api/src/errors/index.ts
@@ -1,3 +1,0 @@
-import { NotFoundError } from './errors'
-
-export { NotFoundError }
