+++ services/app-api/src/postgres/index.ts
@@ -3,5 +3,5 @@ export { InsertHealthPlanPackageArgsType } from './healthPlanPackage'
 export { InsertUserArgsType } from './user'
 export { NewPostgresStore, Store } from './postgresStore'
 export { NewPrismaClient } from './prismaClient'
-export { isStoreError, StoreError } from './storeError'
+export { isStoreError, StoreError, NotFoundError } from './storeError'
 export { findStatePrograms } from './state/findStatePrograms'
